fix: apply suggestion
