--- app/env-list.ts
@@ -1,8 +1,0 @@
-export const ENV_LIST = [
-  'VITE_APP_TIPS_ADDRESS',
-  'VITE_APP_BANK_IMPL_ADDRESS',
-  'VITE_APP_BACKEND_URL',
-  'VITE_APP_VOTING_IMPL_ADDRESS',
-  'VITE_APP_BANK_BEACON_ADDRESS',
-  'VITE_APP_VOTING_BEACON_ADDRESS'
-]
