+++ app/vite.config.ts
@@ -3,8 +3,17 @@ import { fileURLToPath, URL } from 'node:url'
 import { defineConfig, loadEnv } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import inject from '@rollup/plugin-inject'
-import { ENV_LIST } from './env-list'
 import { isAddress } from 'ethers'
+
+export const ENV_LIST = [
+  'VITE_APP_TIPS_ADDRESS',
+  'VITE_APP_BANK_IMPL_ADDRESS',
+  'VITE_APP_BACKEND_URL',
+  'VITE_APP_VOTING_IMPL_ADDRESS',
+  'VITE_APP_BANK_BEACON_ADDRESS',
+  'VITE_APP_VOTING_BEACON_ADDRESS'
+]
+
 
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
