+++ app/vitest.config.ts
@@ -11,7 +11,7 @@ export default defineConfig((env) =>
   mergeConfig(viteConfig(env), {
     test: {
       environment: 'jsdom',
-      exclude: [...configDefaults.exclude, 'test/e2e/*', './env-list'],
+      exclude: [...configDefaults.exclude, 'test/e2e/*'],
       root: fileURLToPath(new URL('./', import.meta.url)),
       coverage: {
         provider: 'istanbul',
